better error handling

+++ controllers/runs.js
@@ -19,24 +19,29 @@ const logger = require('../config/winston')
 module.exports = {
   async getRun (req, res, next) {
     const meta = JSON.parse(await readFile(path.join(getRunsDirPath(req.params.id), 'meta.json')))
-    res.send({ ...meta, ...STATUS.ok })
+    res.send(meta)
   },
 
-  async postRun (req, res) {
+  async postRun (req, res, next) {
     const { submission, test } = req.query
+    const submissionDir = getSubmissionsDirPath(submission)
+    const testDir = getTestsDirPath(test)
+    const testExists = await pathExists(testDir)
+    const { lang, statusCode } = JSON.parse(
+      await readFile(path.join(submissionDir, 'meta.json'))
+    )
+    if (!testExists || !(statusCode === 0)) {
+      return next()
+    }
     const id = createHash(submission + test)
     const taskDir = getTasksDirPath(id)
     const runDir = getRunsDirPath(id)
     const exists = await pathExists(runDir)
     if (exists) {
       logger.info('Duplicate run')
-      return res.send({ id, ...STATUS.ok })
+      return res.send({ id, ...STATUS.queue })
     }
-    const submissionDir = getSubmissionsDirPath(submission)
     await Promise.all([mkdirp(runDir), mkdirp(taskDir)])
-    const { lang } = JSON.parse(
-      await readFile(path.join(submissionDir, 'meta.json'))
-    )
     await Promise.all([
       writeFile(
         path.join(taskDir, 'meta.json'),
@@ -64,7 +69,7 @@ module.exports = {
         exclude: ['meta.json']
       }),
       copyFiles({
-        src: getTestsDirPath(test),
+        src: testDir,
         dst: taskDir,
         exclude: ['meta.json']
       })
